+++ bmc-email/src/main/java/com/oracle/bmc/email/EmailAsync.java
@@ -37,6 +37,22 @@ public interface EmailAsync extends AutoCloseable {
     void setRegion(String regionId);
 
     /**
+     * Moves a sender into a different compartment. When provided, If-Match is checked against ETag values of the resource.
+     *
+     * @param request The request object containing the details to send
+     * @param handler The request handler to invoke upon completion, may be null.
+     * @return A Future that can be used to get the response if no AsyncHandler was
+     *         provided. Note, if you provide an AsyncHandler and use the Future, some
+     *         types of responses (like java.io.InputStream) may not be able to be read in
+     *         both places as the underlying stream may only be consumed once.
+     */
+    java.util.concurrent.Future<ChangeSenderCompartmentResponse> changeSenderCompartment(
+            ChangeSenderCompartmentRequest request,
+            com.oracle.bmc.responses.AsyncHandler<
+                            ChangeSenderCompartmentRequest, ChangeSenderCompartmentResponse>
+                    handler);
+
+    /**
      * Creates a sender for a tenancy in a given compartment.
      *
      * @param request The request object containing the details to send